requirements.txt
python-dotenv
pyairtable
google-generativeai
tenacity